foodgram/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
dateparser==1.1.8
defusedxml==0.7.1
Django==3.2.16
django-cors-headers==4.2.0
django-templated-mail==1.1.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
Faker==19.6.2
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
lark==1.1.7
MarkupSafe==2.1.3
oauthlib==3.2.2
pbr==5.11.1
psycopg2-binary==2.9.9
Pillow==9.2.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
recipe==0.37.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
SQLAlchemy==1.4.49
sqlparse==0.4.4
stevedore==5.1.0
structlog==23.1.0
sureberus==0.15.0
tablib==3.5.0
typing_extensions==4.8.0
tzlocal==5.1
uritemplate==4.1.1
urllib3==2.0.6
users==1.0.dev0

